refactor - separate header from CreateForm (add FirstForm)
